Last
Delete all unecesary clases  the program is fully working,

Error handling is done by try catch and the repositorys are just other clases where I put the DataBase Logic.

I want to make a better error handling more consice and a upgrade my really basic repository to a generic repository and work with the data just in memory allowing the program to work a lot faster.